Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -47,7 +47,7 @@ end
 def get_average_age_for_season(data, season)
   age_array = []
   data[season].each do |contestant|
-    age = contestant["age"].to_i
+    age = contestant["age"].to_f
     age_array << age
   end
   return (age_array.inject(0) {|sum, i|  sum + i } / age_array.size).round
